+++ src/main/java/main/view/publicApi/PublicSuiteServlet.java
@@ -2,7 +2,7 @@ package main.view.publicApi;
 
 import main.Session;
 import main.exceptions.AqualityParametersException;
-import main.model.dto.TestDto;
+import main.model.dto.TestSuiteDto;
 import main.view.BaseServlet;
 import main.view.IPost;
 
@@ -10,8 +10,8 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/public/test/create-or-update")
-public class publicTestServlet extends BaseServlet implements IPost {
+@WebServlet("/public/suite/create-or-update")
+public class PublicSuiteServlet extends BaseServlet implements IPost {
 
     @Override
     public void doPost(HttpServletRequest req, HttpServletResponse resp) {
@@ -21,28 +21,22 @@ public class publicTestServlet extends BaseServlet implements IPost {
             Session session = createSession(req);
             String requestedJson = getRequestJson(req);
 
-            TestDto test = mapper.mapObject(TestDto.class, requestedJson);
+            TestSuiteDto testSuite = mapper.mapObject(TestSuiteDto.class, requestedJson);
+            validatePost(testSuite);
+            testSuite = session.controllerFactory.getHandler(testSuite).createOrUpdate(testSuite);
 
-            validatePost(test);
-            test = session.controllerFactory.getHandler(test).createOrUpdate(test);
-
-            setResponseBody(resp, test);
+            setResponseBody(resp, testSuite);
         } catch (Exception e) {
             handleException(resp, e);
         }
     }
 
-    private void validatePost(TestDto test) throws AqualityParametersException {
-        if(test.getProject_id() == null) {
+    private void validatePost(TestSuiteDto testSuite) throws AqualityParametersException {
+        if(testSuite.getProject_id() == null) {
             throw new AqualityParametersException("You should specify 'project_id'!");
         }
-
-        if(test.getId() == null && test.getName() == null) {
+        if(testSuite.getId() == null && testSuite.getName() == null) {
             throw new AqualityParametersException("You should specify 'id' or/and 'name' suite parameters!");
-        }
-
-        if(test.getSuites() == null) {
-            throw new AqualityParametersException("You should specify 'suites' array with single suite like `[{id: test_suite_id}]`");
         }
     }
